+++ src/components/Uploader/index.js
@@ -1,8 +1,6 @@
 import { h, Component } from 'preact'
 import Dropzone from 'react-dropzone'
-import { DocumentNotFound } from '../Document'
 import Spinner from '../Spinner'
-import Confirm from '../Confirm'
 import theme from '../Theme/style.css'
 import style from './style.css'
 import {functionalSwitch, impurify} from '../utils'
@@ -19,29 +17,57 @@ const UploadProcessing = () =>
     <div className={style.processing}>Processing your document</div>
   </div>
 
-export const UploadError = ({children}) =>
+const UploadError = ({children}) =>
   <div className={`${style.text} ${style.error}`}>{children}</div>
 
-const InvalidCaptureError = ({method}) => functionalSwitch(method, {
-  document: () => <DocumentNotFound />
-})
+const InvalidCapture = ({message}) =>
+  <UploadError>{message}</UploadError>
 
-const InvalidFileType = () =>
-  <UploadError>The file uploaded has an unsupported file type.</UploadError>
+InvalidCapture.defaultProps = {
+  message: 'We are unable to detect an identity document in this image. Please try again.'
+}
+
+const InvalidFileType = ({message}) =>
+  <UploadError>{message}</UploadError>
+
+InvalidFileType.defaultProps = {
+  message: 'The file uploaded has an unsupported file type.'
+}
+
+const InvalidFileSize = ({message}) =>
+  <UploadError>{message}</UploadError>
+
+const ServerError = ({message}) =>
+  <UploadError>{message}</UploadError>
+
+InvalidFileSize.defaultProps = {
+  message: 'The file size limit of 4MB has been exceeded. Please try again.'
+}
+
+ServerError.defaultProps = {
+  message: 'There was an error connecting to the server. Please wait and try again later.'
+}
 
 //TODO move to react instead of preact, since preact has issues handling pure components
 //IF this component is exported as pure,
 //some components like Camera will not have componentWillUnmount called
 export const Uploader = impurify(({method, onImageSelected, uploading, error}) => (
   <Dropzone
-    onDrop={([ file ])=> onImageSelected(file)}
+    onDrop={([ file ])=> {
+      //removes a memory leak created by react-dropzone
+      URL.revokeObjectURL(file.preview)
+      delete file.preview
+      onImageSelected(file)
+    }}
     multiple={false}
     className={style.dropzone}
   >
     {uploading ? <UploadProcessing /> : <UploadInstructions />}
     {!uploading && functionalSwitch(error, {
-      INVALID_CAPTURE: ()=> <InvalidCaptureError {...{method}}/>,
-      INVALID_TYPE: ()=>    <InvalidFileType />
+      INVALID_CAPTURE: () => <InvalidCapture />,
+      INVALID_TYPE: () => <InvalidFileType />,
+      INVALID_SIZE: () => <InvalidFileSize />,
+      SERVER_ERROR: () => <ServerError />
     })}
   </Dropzone>
 ))