+++ src/components/utils/file.js
@@ -1,16 +1,31 @@
 import loadImage from 'blueimp-load-image/js/load-image'
+import 'blueimp-load-image/js/load-image-orientation'
+import 'blueimp-load-image/js/load-image-exif'
 import {canvasToBase64Images, toLossyImageDataUrl} from './canvas.js'
 
 export const fileToBase64 = (file, callback, errorCallback) => {
-  const reader = new FileReader();
-  reader.readAsDataURL(file);
-  reader.onload = function () {
-    callback(reader.result)
-  };
+  const reader = new FileReader()
+  reader.readAsDataURL(file)
+  reader.onload = () => {callback(reader.result)}
   reader.onerror = function (error) {
-    console.log('File Reading Error: ', error);
-    errorCallback(error)
-  };
+   console.log('File Reading Error: ', error)
+   errorCallback(error)
+ }
+}
+
+const decodeBase64 = (image) => {
+  const byteString  = atob(image.split(',')[1])
+  const mimeString = image.split(',')[0].split(':')[1].split(';')[0]
+  const integerArray = new Uint8Array(byteString.length)
+  for (let i = 0; i < byteString.length; i++) {
+    integerArray[i] = byteString.charCodeAt(i)
+  }
+  return {integerArray, mimeString}
+}
+
+export const base64toBlob = (image) => {
+  const base64Data = decodeBase64(image)
+  return new Blob([base64Data.integerArray], {type: base64Data.mimeString})
 }
 
 export const fileType = file => file.type.split('/')[1]
@@ -19,19 +34,18 @@ export const isOfFileType = (fileTypeList, file) =>
   fileTypeList.some(acceptableFileType =>
     acceptableFileType === fileType(file));
 
-const fileToCanvas = (options = { maxWidth: 960, maxHeight: 960, canvas: true},
-                      file, callback, errorCallback) =>
-  loadImage(file.preview, canvasOrEventError => {
+const fileToCanvas = (file, callback, errorCallback) =>
+  loadImage(file, canvasOrEventError => {
     if (canvasOrEventError.type === "error"){
       errorCallback(canvasOrEventError)
     }
     else {
       callback(canvasOrEventError)
     }
-  }, options)
+  }, { maxWidth: 960, maxHeight: 960, orientation: true })
 
-export const fileToLossyBase64Image = (options, file, callback, errorCallback) =>
-  fileToCanvas(options, file,
+export const fileToLossyBase64Image = (file, callback, errorCallback) =>
+  fileToCanvas(file,
     canvas => toLossyImageDataUrl(canvas, callback),
     errorCallback
   )