+++ src/index.js
@@ -1,44 +1,24 @@
 import { h, render, Component } from 'preact'
 import { Provider } from 'react-redux'
-import { store, events, connect as ws } from 'onfido-sdk-core'
+import { store, events } from 'onfido-sdk-core'
 import Modal from './components/Modal'
-import AppRouter from './components/Router'
+import Router from './components/Router'
 import forEach from 'object-loops/for-each'
 import mapValues from 'object-loops/map'
 import Tracker from './Tracker'
 
 Tracker.setUp()
 
-
-
 const ModalApp = ({ options:{ useModal, isModalOpen, buttonId, ...otherOptions},
                     ...otherProps}) =>
   <Modal {...{useModal, buttonId}} isOpen={isModalOpen}>
-    <AppRouter options={otherOptions} {...otherProps}/>
+    <Router options={otherOptions} {...otherProps}/>
   </Modal>
 
-const ContainerPure = ({ options, socket }) =>
+const Container = ({ options }) =>
   <Provider store={store}>
-    <ModalApp options={options} socket={socket}/>
+    <ModalApp options={options}/>
   </Provider>
-
-class Container extends Component {
-  componentWillMount () {
-    const { token, socketUrl } = this.props.options
-    this.setState({ socket:ws(token, socketUrl) })
-  }
-
-  componentWillReceiveProps (nextProps) {
-    const { token: nextToken, socketUrl: nextSocketUrl } = nextProps.options
-    const { token, socketUrl } = this.props.options
-    if (token !== nextToken || socketUrl !== nextSocketUrl) {
-      this.setState({ socket:ws(nextToken, nextSocketUrl) })
-    }
-  }
-
-  render = ({options}) =>
-    <ContainerPure {...this.props} socket={this.state.socket}/>
-}
 
 /**
  * Renders the Onfido component
@@ -50,27 +30,28 @@ const onfidoRender = (options, el, merge) => {
   return render( <Container options={options}/>, el, merge)
 }
 
+const stripOneCapture = ({blob, documentType, id, side}) => {
+  const capture = {id, blob}
+  if (documentType) capture.documentType = documentType
+  if (side) capture.side = side
+  return capture
+}
 
-const stripOneCapture = ({image,documentType,id}) =>
-  documentType === undefined ? {id,image} : {id,image,documentType}
-
-const stripCapturesHashToNecessaryValues = captures => mapValues(captures,
+const stripCapturesHash = captures => mapValues(captures,
   capture => capture ? stripOneCapture(capture) : null)
 
+const getCaptures = () => stripCapturesHash(events.getCaptures())
+
 function bindEvents (options) {
-  const strip = stripCapturesHashToNecessaryValues
   const eventListenersMap = {
-    ready: () => { options.onReady() },
-    documentCapture: data => { options.onDocumentCapture(stripOneCapture(data)) },
-    faceCapture: data => { options.onFaceCapture(stripOneCapture(data)) },
-    complete: data => { options.onComplete(strip(data)) }
+    ready: () => options.onReady(),
+    documentCapture: () => options.onDocumentCapture(getCaptures().documentCapture),
+    documentBackCapture: () => options.onDocumentCapture(getCaptures().documentBackCapture),
+    faceCapture: () => options.onFaceCapture(getCaptures().faceCapture),
+    complete: () => options.onComplete(getCaptures())
   }
 
-  forEach(eventListenersMap, (listener, event) => {
-    if (event === 'ready') events.once(event, listener)
-    else events.on(event, listener)
-  })
-
+  forEach(eventListenersMap, (listener, event) => events.on(event, listener))
   return eventListenersMap;
 }
 
@@ -88,7 +69,7 @@ function rebindEvents(newOptions, previousEventListenersMap){
 
 const Onfido = {}
 
-Onfido.getCaptures = () => stripCapturesHashToNecessaryValues(events.getCaptures())
+Onfido.getCaptures = () => getCaptures()
 
 const noOp = ()=>{}