+++ CHANGELOG.md
@@ -9,10 +9,22 @@ Breaking changes result in a different major. UI changes that might break custom
 ### Added
 - Internal: added blur/clarity client-side detection when autoCapture is active
 
+## [0.14.0]
 
-## [0.13.0-rc.2]
+### Changed
 
-Install with `npm install onfido-sdk-ui@0.13.0-rc.2`
+- Public: Document and face captures will be returned by callbacks as File or Blob instead of base64.
+- Internal: Callbacks are now linked to the flow rather than the Redux store.
+
+### Added
+
+- Public: Capture the reverse side of driving licenses and ID cards.
+- Public: Add a file size limit of 4 MB in line with the Onfido API.
+
+### Fixed
+- Internal: Read exif tags to orientate images correctly.
+
+## [0.13.0]
 
 ### Changed
 
@@ -166,8 +178,9 @@ Install with `npm install onfido-sdk-ui@0.12.0-rc.1`
 
 The standard for this change log can be found [here](http://keepachangelog.com/).
 
-[next-version]: https://github.com/onfido/onfido-sdk-ui/compare/0.13.0-rc.2...master
-[0.13.0-rc.1]: https://github.com/onfido/onfido-sdk-ui/compare/0.12.0-rc.1...0.13.0-rc.2
+[next-version]: https://github.com/onfido/onfido-sdk-ui/compare/0.14.0...master
+[0.14.0]: https://github.com/onfido/onfido-sdk-ui/compare/0.13.0...0.14.0
+[0.13.0]: https://github.com/onfido/onfido-sdk-ui/compare/0.12.0-rc.1...0.13.0
 [0.12.0-rc.1]: https://github.com/onfido/onfido-sdk-ui/compare/0.11.1...0.12.0-rc.1
 [0.11.1]: https://github.com/onfido/onfido-sdk-ui/compare/0.11.0...0.11.1
 [0.11.0]: https://github.com/onfido/onfido-sdk-ui/compare/0.10.0...0.11.0